Move graph representation into dependency tree
